+++ drake/systems/plants/inverseKinBackend.cpp
@@ -179,7 +179,7 @@ void inverseKinMode1(
 
         typedef Eigen::Triplet<double> T;
         std::vector<T> triplet_list;
-        for (size_t i = 0; i < iAfun.size(); i++) {
+        for (int i = 0; i < iAfun.size(); i++) {
           triplet_list.push_back(T(iAfun[i], jAvar[i], A[i]));
         }
